Refactor db
