+++ mirar/database/__init__.py
@@ -1,0 +1,3 @@
+"""
+Central module for database-related functions
+"""
