+++ mirar/database/constants.py
@@ -1,0 +1,6 @@
+"""
+Central list of postgres constants
+"""
+
+
+POSTGRES_DUPLICATE_PROTOCOLS = ["fail", "ignore", "replace"]
