+++ mirar/database/user/__init__.py
@@ -1,0 +1,11 @@
+"""
+Module containing classes for interacting with postgres database
+"""
+from mirar.database.user.credentials import (
+    ADMIN_PASSWORD,
+    ADMIN_USER,
+    DB_PASSWORD,
+    DB_USER,
+)
+from mirar.database.user.postgres_admin import PostgresAdmin
+from mirar.database.user.postgres_user import PostgresUser
