+++ mirar/pipelines/winter/models/_astrometry_stats.py
@@ -8,9 +8,9 @@ from pydantic import Field
 from sqlalchemy import Column, Float, ForeignKey  # event,
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
+from mirar.database.base_model import BaseDB, dec_field, ra_field
 from mirar.pipelines.winter.models._raw import RawTable
 from mirar.pipelines.winter.models.base_model import WinterBase
-from mirar.processors.sqldatabase.base_model import BaseDB, dec_field, ra_field
 
 # from mirar.utils.sql import create_q3c_extension
