+++ mirar/pipelines/winter/models/_ref_queries.py
@@ -7,8 +7,8 @@ from pydantic import Field
 from sqlalchemy import VARCHAR, Column, Float, ForeignKey, Integer
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
+from mirar.database.base_model import BaseDB
 from mirar.pipelines.winter.models.base_model import WinterBase, dec_field, ra_field
-from mirar.processors.sqldatabase.base_model import BaseDB
 
 
 class RefQueriesTable(WinterBase):
