+++ mirar/pipelines/winter/models/_ref_stacks.py
@@ -6,8 +6,8 @@ from typing import ClassVar
 from pydantic import Field
 from sqlalchemy import VARCHAR, Column, Double, Float, Integer
 
+from mirar.database.base_model import BaseDB
 from mirar.pipelines.winter.models.base_model import WinterBase, dec_field, ra_field
-from mirar.processors.sqldatabase.base_model import BaseDB
 
 
 class RefStacksTable(WinterBase):
