+++ mirar/pipelines/winter/models/_subdets.py
@@ -7,10 +7,10 @@ from pydantic import Field
 from sqlalchemy import Column, Integer, Select
 from sqlalchemy.orm import Mapped, relationship
 
+from mirar.database.base_model import BaseDB, _exists
+from mirar.database.engine import get_engine
 from mirar.pipelines.winter.constants import subdets
 from mirar.pipelines.winter.models.base_model import WinterBase
-from mirar.processors.sqldatabase.base_model import BaseDB, _exists
-from mirar.utils.sql import get_engine
 
 DEFAULT_FIELD = 999999999
