+++ mirar/pipelines/winter/models/base_model.py
@@ -6,7 +6,7 @@ from datetime import date
 from pydantic import Field
 from sqlalchemy.orm import DeclarativeBase
 
-from mirar.processors.sqldatabase.base_model import BaseTable
+from mirar.database.base_model import BaseTable
 
 DB_NAME = "winter"
